Added test cases for QuadTree exception cases

+++ QuadTreeDemo/QuadTreeDemo.cs
@@ -161,11 +161,11 @@ namespace QuadTreeTest
             var simpleCount = 0;
             Clock.BenchmarkTime(() =>
             {
-                m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
-                m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
-                m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
-                m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
-                m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
+                m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
+                m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
+                m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
+                m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
+                m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true));
             }, 10000);
 
             Console.WriteLine("End Simple Test, checked: " + simpleCount);
@@ -199,11 +199,11 @@ namespace QuadTreeTest
                 threads.Add(new Thread(() => { m_Tree.GetClosestObject(GetRandomPos(), (float) (100 + Random.NextDouble() * 400)); }));
                 threads.Add(new Thread(() => { m_Tree.GetClosestObject(GetRandomPos(), (float) (100 + Random.NextDouble() * 400)); }));
 
-                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
-                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
-                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
-                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
-                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
+                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
+                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
+                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
+                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
+                threads.Add(new Thread(() => { m_Tree.GetKClosestObjects(GetRandomPos(), (uint) Random.Next(5, 20), (float)(100 + Random.NextDouble() * 400), new PriorityQueue<Transformable>(true)); }));
 
                 threads.Add(new Thread(() => { m_Tree.GetObjectsInRange(GetRandomPos(), (float)(100 + Random.NextDouble() * 400), new List<Transformable>()); }));
                 threads.Add(new Thread(() => { m_Tree.GetObjectsInRange(GetRandomPos(), (float)(100 + Random.NextDouble() * 400), new List<Transformable>()); }));
